accounts/abi: separation of the testing suite into appropriately named files.
